Rework dataset collect script, update dataset and add requirements.txt
