+++ requirements.txt
@@ -1,0 +1,5 @@
+pandas~=2.3.3
+numpy~=2.3.3
+matplotlib~=3.10.7
+seaborn~=0.13.2
+requests~=2.32.5
